Add Current account form and fix UI ManagementForm

+++ PRN211-BirdFarmShop/StaffManagerForm.cs
@@ -13,12 +13,14 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Xml.Linq;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.ListView;
 
 namespace PRN211_BirdFarmShop
 {
     public partial class StaffManagerForm : Form
     {
         private IAccountService accountService;
+        public String CurrentEmailAccount;
         public StaffManagerForm()
         {
             InitializeComponent();
@@ -33,7 +35,7 @@ namespace PRN211_BirdFarmShop
                 txtStreet.Text.Trim().Length == 0 ||
                 txtPhone.Text.Trim().Length == 0 ||
                 dt_birthday.Text.Trim().Length == 0 ||
-                (!radioMale.Checked && !radioMale.Checked))
+                (!radioMale.Checked && !radioFemale.Checked))
             {
                 return false;
             }
@@ -54,85 +56,7 @@ namespace PRN211_BirdFarmShop
             }
         }
 
-        void UpdateFont()
-        {
-            var colIndex = 0;
-            //Change cell font
-            foreach (DataGridViewColumn c in dtg_staffList.Columns)
-            {
-                c.DefaultCellStyle.Font = new Font("Microsoft Sans Serif", 18F, GraphicsUnit.Pixel);
-                // header font custom
-                dtg_staffList.Columns[colIndex].HeaderCell.Style.Font = new Font("Tahoma", 9.75F, FontStyle.Bold);
 
-                // increase column index 
-                ++colIndex;
-            }
-        }
-
-        /// <summary>
-        /// Styling form
-        /// </summary>
-        void StyleDatagridview()
-        {
-
-            dtg_staffList.BorderStyle = BorderStyle.None;
-            dtg_staffList.AlternatingRowsDefaultCellStyle.BackColor = Color.FromArgb(238, 239, 249);
-            dtg_staffList.CellBorderStyle = DataGridViewCellBorderStyle.SingleHorizontal;
-            dtg_staffList.DefaultCellStyle.SelectionBackColor = Color.SeaGreen;
-            dtg_staffList.DefaultCellStyle.SelectionForeColor = Color.WhiteSmoke;
-            //dgv_Category.BackgroundColor = Color.FromArgb(240, 240, 240);
-            dtg_staffList.RowHeadersWidthSizeMode = DataGridViewRowHeadersWidthSizeMode.DisableResizing;//optional
-            dtg_staffList.EnableHeadersVisualStyles = false;
-            dtg_staffList.ColumnHeadersBorderStyle = DataGridViewHeaderBorderStyle.None;
-            dtg_staffList.ColumnHeadersDefaultCellStyle.Font = new Font("MS Reference Sans Serif", 10);
-            dtg_staffList.ColumnHeadersDefaultCellStyle.BackColor = Color.FromArgb(18, 122, 219);
-            dtg_staffList.ColumnHeadersDefaultCellStyle.ForeColor = Color.White;
-        }
-
-        /// <summary>
-        /// resize form when shrink
-        /// </summary>
-        void ResizeForm()
-        {
-            // no smaller than design time size
-            //this.MinimumSize = new System.Drawing.Size(this.Width, this.Height);
-
-            // no larger than screen size
-            //this.MaximumSize = new Size(Screen.PrimaryScreen.Bounds.Width, Screen.PrimaryScreen.Bounds.Height, 0);
-
-            // get all form elements
-            var controls = GetAllControls(this).OfType<Control>();
-            // resize form element  
-            foreach (var control in controls)
-            {
-                Resize_Control(control, new Rectangle(control.Location, control.Size));
-            }
-
-            Resize_Control(bt_add, new Rectangle(bt_add.Location, bt_add.Size));
-
-            //this.AutoSize = true;
-            //this.AutoSizeMode = AutoSizeMode.GrowAndShrink;
-        }
-
-        /// <summary>
-        /// Resize form controls
-        /// </summary>
-        /// <param name="c"></param>
-        /// <param name="r"></param>
-        void Resize_Control(Control c, Rectangle r)
-        {
-            float xRatio = (float)(this.Width) / (float)(this.Size.Width);
-            float yRatio = (float)(this.Height) / (float)(this.Size.Height);
-            int newX = (int)(r.X * xRatio);
-            int newY = (int)(r.Y * yRatio);
-
-            int newWidth = (int)(r.Width * xRatio);
-            int newHeight = (int)(r.Height * yRatio);
-
-            c.Location = new Point(newX, newY);
-            c.Size = new Size(newWidth, newHeight);
-
-        }
 
         public bool chkManager(String email)
         {
@@ -166,7 +90,8 @@ namespace PRN211_BirdFarmShop
 
         private void StaffManagerForm_Load(object sender, EventArgs e)
         {
-            StyleDatagridview();
+            txtEmail.Enabled = true;
+            //StyleDatagridview();
             //FormatListStaff();
             var stafflist = accountService.getAllAccountByRole(Constants.IsStaff).
                      Select(staff => new
@@ -185,47 +110,30 @@ namespace PRN211_BirdFarmShop
             // set to bs datasoure <- categories
             bindingSource.DataSource = stafflist;
             dtg_staffList.DataSource = bindingSource;
-            if (dtg_staffList.Columns.Count > 1)
-            {
-                // update to dgv datasource
-                dtg_staffList.DataSource = bindingSource;
-                dtg_staffList.Columns["Fullname"].HeaderText = "Họ và tên";
-                dtg_staffList.Columns["Street"].HeaderText = "Đường";
-                dtg_staffList.Columns["City"].HeaderText = "Thành Phố";
-                dtg_staffList.Columns["Phone"].HeaderText = "SĐT";
-                dtg_staffList.Columns["Street"].HeaderText = "Đường";
-                dtg_staffList.Columns["Birthday"].HeaderText = "Ngày Sinh";
-                dtg_staffList.Columns["Gender"].HeaderText = "Giới tính";
-                dtg_staffList.Columns["RoleDesc"].HeaderText = "Chức vụ";
 
-                dtg_staffList.Columns[0].Width = 190;
-                dtg_staffList.Columns[1].Width = 190;
-                dtg_staffList.Columns[2].Width = 130;
-                dtg_staffList.Columns[7].Width = 70;
-            }
-            // custom data grid view column
-            CustomDataGridViewColumn();
-
-            // custom font
-            UpdateFont();
         }
 
         private void dtg_staffList_CellClick(object sender, DataGridViewCellEventArgs e)
         {
-            txtEmail.Text = dtg_staffList.CurrentRow.Cells["Email"].Value.ToString();
-            txtFullname.Text = dtg_staffList.CurrentRow.Cells["Fullname"].Value.ToString();
-            txtStreet.Text = dtg_staffList.CurrentRow.Cells["Street"].Value.ToString();
-            txtCity.Text = dtg_staffList.CurrentRow.Cells["City"].Value.ToString();
-            txtPhone.Text = dtg_staffList.CurrentRow.Cells["Phone"].Value.ToString();
-            dt_birthday.Text = dtg_staffList.CurrentRow.Cells["Birthday"].Value.ToString();
-            string gender = dtg_staffList.CurrentRow.Cells["Gender"].Value.ToString();
-            if (gender.ToLower() == "nam")
+            var row = dtg_staffList.Rows.Count;
+
+            if (e.RowIndex != -1 && e.RowIndex < row)
             {
-                radioMale.Checked = true;
-            }
-            else
-            {
-                radioFemale.Checked = true;
+                txtEmail.Text = dtg_staffList.CurrentRow.Cells["Email"].Value.ToString();
+                txtFullname.Text = dtg_staffList.CurrentRow.Cells["Fullname"].Value.ToString();
+                txtStreet.Text = dtg_staffList.CurrentRow.Cells["Street"].Value.ToString();
+                txtCity.Text = dtg_staffList.CurrentRow.Cells["City"].Value.ToString();
+                txtPhone.Text = dtg_staffList.CurrentRow.Cells["Phone"].Value.ToString();
+                dt_birthday.Text = dtg_staffList.CurrentRow.Cells["Birthday"].Value.ToString();
+                string gender = dtg_staffList.CurrentRow.Cells["Gender"].Value.ToString();
+                if (gender.ToLower() == "nam")
+                {
+                    radioMale.Checked = true;
+                }
+                else
+                {
+                    radioFemale.Checked = true;
+                }///
             }
 
         }
@@ -290,7 +198,6 @@ namespace PRN211_BirdFarmShop
                     return;
                 }
 
-
                 Account account = new Account();
                 account.Email = txtEmail.Text;
                 account.Fullname = txtFullname.Text;
@@ -318,7 +225,11 @@ namespace PRN211_BirdFarmShop
 
                 if (isSucess)
                 {
-                    MessageBox.Show("Tạo mới Nhân Viên thành công");
+                    MessageBox.Show("Tạo mới Nhân Viên thành công \n" +
+                        "Nhân viên:" + txtFullname.Text + "\n" +
+                        "Có tài khoảng đăng nhập: " + txtEmail.Text + "\n" +
+                        "Mật khẩu mặc định: 123456");
+
 
                     // reset features
                     button6_Click(sender, e);
@@ -358,7 +269,6 @@ namespace PRN211_BirdFarmShop
             try
             {
                 var email = txtEmail.Text.Trim();
-
                 // get category by id
                 var existEmail = accountService.GetAccount(email);
                 if (existEmail is null)
@@ -366,17 +276,20 @@ namespace PRN211_BirdFarmShop
                     MessageBox.Show("Không tìm thấy nhân viên");
                     return;
                 }
+                if (MessageBox.Show("You really want to delete this staff!", "Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
+                {
 
-                // delete category by id
-                accountService.Delete(email);
+                    // delete category by id
+                    accountService.Delete(email);
 
-                // load form data
-                StaffManagerForm_Load(sender, e);
+                    // load form data
+                    StaffManagerForm_Load(sender, e);
 
-                MessageBox.Show("Xóa nhân viên thành công");
+                    MessageBox.Show("Xóa nhân viên thành công");
 
-                // reset features
-                button6_Click(sender, e);
+                    // reset features
+                    button6_Click(sender, e);
+                }
             }
             catch (Exception ex)
             {
@@ -432,39 +345,40 @@ namespace PRN211_BirdFarmShop
                     return;
                 }
 
-
-
-                existEmail.Email = txtEmail.Text;
-                existEmail.Fullname = txtFullname.Text;
-                existEmail.City = txtCity.Text;
-                existEmail.Street = txtStreet.Text;
-                existEmail.RoleId = 2;
-                existEmail.Password = "123456";
-                existEmail.Birthday = DateTime.Parse(dt_birthday.Text);
-                existEmail.Phone = txtPhone.Text;
-                if (radioFemale.Checked)
+                if (MessageBox.Show("Bạn chắc chắn muốn cập nhật tài khoản này!", "Cảnh Báo", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning) == DialogResult.OK)
                 {
-                    existEmail.Gender = radioFemale.Text;
-                }
-                else
-                {
-                    existEmail.Gender = radioMale.Text;
-                }
-                // create new category
-                bool isSucess = accountService.Update(Email, existEmail);
 
-                if (isSucess)
-                {
-                    MessageBox.Show("Cập nhật thông tin Nhân Viên thành công\n" +
-                        "Nhân viên mới sẽ có Email là:" + existEmail.Email + "\n" +
-                        "Mật khẩu mặc định: 123456");
 
-                    // reset features
-                    button6_Click(sender, e);
+                    existEmail.Email = txtEmail.Text;
+                    existEmail.Fullname = txtFullname.Text;
+                    existEmail.City = txtCity.Text;
+                    existEmail.Street = txtStreet.Text;
+                    existEmail.RoleId = 2;
+                    //existEmail.Password = "123456";
+                    existEmail.Birthday = DateTime.Parse(dt_birthday.Text);
+                    existEmail.Phone = txtPhone.Text;
+                    if (radioFemale.Checked)
+                    {
+                        existEmail.Gender = radioFemale.Text;
+                    }
+                    if (!radioFemale.Checked)
+                    {
+                        existEmail.Gender = radioMale.Text;
+                    }
+                    // create new category
+                    bool isSucess = accountService.Update(Email, existEmail);
 
-                    // load form data
-                    StaffManagerForm_Load(sender, e);
-                    return;
+                    if (isSucess)
+                    {
+                        MessageBox.Show("Cập nhật thông tin Nhân Viên thành công\n");
+
+                        // reset features
+                        button6_Click(sender, e);
+
+                        // load form data
+                        StaffManagerForm_Load(sender, e);
+                        return;
+                    }
                 }
 
                 MessageBox.Show("Cập nhật thông tin Nhân Viên thất bại");
@@ -477,7 +391,6 @@ namespace PRN211_BirdFarmShop
 
         private void txtSearch_TextChanged(object sender, EventArgs e)
         {
-            StyleDatagridview();
             var searchvalue = txtSearch.Text.Trim();
             var stafflist = accountService.SearchAccountInRole(searchvalue, Constants.IsStaff).
                      Select(staff => new
@@ -495,39 +408,23 @@ namespace PRN211_BirdFarmShop
             // set to bs datasoure <- categories
             bindingSource.DataSource = stafflist;
             dtg_staffList.DataSource = bindingSource;
-            if (dtg_staffList.Columns.Count > 1)
-            {
-                // update to dgv datasource
-                dtg_staffList.DataSource = bindingSource;
-                dtg_staffList.Columns["Fullname"].HeaderText = "Họ và tên";
-                dtg_staffList.Columns["Street"].HeaderText = "Đường";
-                dtg_staffList.Columns["City"].HeaderText = "Thành Phố";
-                dtg_staffList.Columns["Phone"].HeaderText = "SĐT";
-                dtg_staffList.Columns["Street"].HeaderText = "Đường";
-                dtg_staffList.Columns["Birthday"].HeaderText = "Ngày Sinh";
-                dtg_staffList.Columns["Gender"].HeaderText = "Giới tính";
-                dtg_staffList.Columns["RoleDesc"].HeaderText = "Chức vụ";
-
-                dtg_staffList.Columns[0].Width = 190;
-                dtg_staffList.Columns[1].Width = 190;
-                dtg_staffList.Columns[2].Width = 130;
-                dtg_staffList.Columns[7].Width = 70;
-            }
-            // custom data grid view column
-            CustomDataGridViewColumn();
-
-            // custom font
-            UpdateFont();
-        }
-
-        private void groupBox2_Enter(object sender, EventArgs e)
-        {
 
         }
 
-        private void groupBox4_Enter(object sender, EventArgs e)
-        {
 
+
+
+        private void bt_close_Click(object sender, EventArgs e)
+        {
+            this.Close();
+            System.Windows.Forms.Application.Exit();
+        }
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            AccountInfoForm accountInfoForm = new AccountInfoForm();
+            accountInfoForm.Show();
+            accountInfoForm.CurrentAccount = accountService.GetAccount(CurrentEmailAccount);
         }
     }
 }
